src/main/java/com/packt/webstore/service/impl/CustomerServiceImpl.java
package com.packt.webstore.service.impl;

import com.packt.webstore.domain.Customer;
import com.packt.webstore.domain.repository.CustomerRepository;
import com.packt.webstore.service.CustomerService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import java.util.List;

/**
 * Created by okeynan on 4/19/15.
 */
@Service
public class CustomerServiceImpl implements CustomerService {

    @Autowired
    CustomerRepository mCustomerRepository;

    @Override
    public List<Customer> getAllCustomers() {
        return mCustomerRepository.getAllCustomers();
    }
}
